entrypoint.sh: Implemented the CI initiative
This is the code used for the DevOps Initiative on GNOME apps, with minor changes. This builds and test the app, and then export a Flatpak bundle. Note: This does not download the app code, instead use the repo from /github/workspace, so the workflow configured for this project will fail