Add a cache-key option
Sometimes, if the action fails to save the cache or if the job was cancelled while it
tries to save the cache, it reserves the cache-key but never free it.

Which causes the ReservedCacheError to be thrown at saveCache step. This allows
the end-user to tweak it to their fiting if needed